+++ ReleaseNotes/release_notes.md
@@ -4,6 +4,8 @@
 
 ## Bug fixes
 
+   1. tiny fix in `Ostap::Math::(Gen)Hyperbolic::setStandard`
+
 # v3.2.0.0 
 
 ## New features